Update page2-sign-up.js
Adding a new className: 'infoForm'

Adding "pattern={formData.newPassword}" to ensure that the two passwords input are identical.

The pattern in Line 128 is irrelevant to the email address web form.

The option "SELECT" is added to the state webform.

The submit button (FINISH) is added to the inputRight div.

inputButtonSection is replaced with inputField.