Update FYP_stylesheet.css
The grid in the inputSection element is removed due to the addition of the inputForm form.

inputTitle is given some padding.

A new grid is included in the new inputForm form.

Grid positions of inputLeft and inputRight are updated.

The new ID "option" is added to center the options in the state webform.