Update masthead-blue.js
Line 1
Imported useState from React

Line 3
Imported the React component SignOutModal.

Line 9
The prop setSignedIn is no longer necessary as this state setter is implemented in SignOutModal through useContext.

Line 13
Defined the state constants modal and setModal for the SignOutModal component.

Red Lines 18 - 20
These lines were moved to the SignOutModal component.

Lines 21 - 22
The function signOutModal() (not to be mistaken with the similarly named component) changes the state of the modal state to true.

Lines 27 and 29
Replaved the function signOut() attached to the "SIGN OUT" button with the signOutModal () function.

Line 32
Set React conditionals for the SignOutModal component.
The component will appear only when the modal state is true, which happens only when signOutModal() function is executed.