Update modal-home.js
Line 1
Imported useContext from React.

Line 2
Imported LoginContext from the login-context.js file.

Lines 3 and 5
Removed the setSignedIn prop due the implementation of useContext and LoginContext.

Line 7
Defined the constant setSignedIn using useContext and LoginContext.

Lines 22, 25 and 26
Replaced the text at Line 22 with
the text at Lines 25 and 26.
The text at these lines are placeholders for web forms which will be implemented in the future.