New Modals and Minor Overhauls.
The modals "FetchErrorModal" and "FetchLoadingModal" are imported to these files.

Some of the import paths were rearranged.

Lines 17 - 24
These are conditionals which will trigger "FetchErrorModal" and "FetchLoadingModal" respectively.

The two new modal component are placed above the main component with conditionals.

all instances of {loading && <p className= "loadingPlaceholder">Loading...</p>} are removed because "FetchLoadingModal" is more efficient.

In each paragraph, a "for" attribute is added, and the "data &&" conditional is removed.

The entire main component is given the conditional "data &&".